# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Test tracking and run management REST API.",
        "title": "TTGO API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/analytics/flaky": {
            "get": {
                "description": "Returns tests ranked by status-switch percentage",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Detect flaky tests (switch method)",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "Number of recent runs to analyze per test",
                        "name": "lookback",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Max results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/analytics/summary": {
            "get": {
                "description": "Returns an aggregated summary of test run analytics",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get analytics summary",
                "responses": {
                    "200": {
                        "description": "Analytics summary object",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/analytics/trend": {
            "get": {
                "description": "Returns time-series trend data for test runs over a given number of days",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get trend data",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "Number of days to include in the trend",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trend points array and days count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: \"Bearer \u003cyour-token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}